Make rank uint32_t. Fix profiler test

+++ tt_metal/hw/inc/accessor/detail/helpers.hpp
@@ -7,7 +7,7 @@
 namespace nd_sharding {
 namespace detail {
 namespace {
-template <template <size_t...> class Wrapper, size_t BASE_IDX, size_t... Is>
+template <template <uint32_t...> class Wrapper, size_t BASE_IDX, size_t... Is>
 constexpr auto make_struct_from_sequence_wrapper(std::index_sequence<Is...>)
     -> Wrapper<get_compile_time_arg_val(BASE_IDX + Is)...>;
 
@@ -23,7 +23,7 @@ constexpr std::array<uint32_t, sizeof...(Is)> make_rta_array_from_sequence(std::
 
 }  // namespace
 
-template <template <size_t...> class Wrapper, size_t base, size_t rank>
+template <template <uint32_t...> class Wrapper, size_t base, uint32_t rank>
 using struct_cta_sequence_wrapper_t =
     decltype(make_struct_from_sequence_wrapper<Wrapper, base>(std::make_index_sequence<rank>{}));
